Configuration splitted
